refactor: extract the panel's remaining leaf components to nav/
ListToolbar (with its six one-off icons), TabStrip, UploadStatusList and
GroupSelectCheckbox move out. Props unchanged; ListToolbar keeps reading
useSelection/useMoveClipboard from context as it already did.

Three zero-behaviour tidies while the code was open:
  - the min-w-0 comment was duplicated verbatim
  - selectAllFor took a whole SectionNode but read only the name and the
    items' ids, which made the call site invent a fake root node
    ({ id: "__root__", children: [] }); it now takes (name, ids)
  - rowList's indentClass default was never overridden by its one caller

Pure move: RecordingsPanel.test.tsx is byte-identical and all 85 cases pass.
Panel 1068 -> 720.
