+++ apps/web/package.json
@@ -1,7 +1,7 @@
 {
   "name": "diariz-web",
   "private": true,
-  "version": "0.186.6",
+  "version": "0.186.7",
   "type": "module",
   "scripts": {
     "dev": "vite",
