+++ version.json
@@ -1,1 +1,1 @@
-{ "version": "0.186.6" }
+{ "version": "0.186.7" }
